backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==3.0.1
gunicorn==21.2.0
nginx
Copy
flask
google-cloud-logging
langchain
google-cloud-aiplatform
faiss-cpu
scikit-learn
joblib
crewai
